requirements.txt
mlx-vlm==0.1.21
fastapi
uvicorn
Pillow